chore: Some change s applied
