+++ package.json
@@ -21,7 +21,7 @@
     "react-bootstrap": "^2.0.0-beta.4",
     "react-bootstrap-table2-paginator": "^2.1.2",
     "react-dom": "^17.0.1",
-    "react-icons": "^4.1.0",
+    "react-icons": "^4.3.1",
     "react-router-dom": "^5.2.0",
     "react-search-autocomplete": "^2.0.0",
     "react-svg-radar-chart": "^1.2.0",
